feat(backend): 调通bsc-mcp并完成BNB Chain业务测试
## 主要改动

### 1. 数据库字段回退
- 回退部分数据库字段修改,使用MySQL作为开发环境数据库
- 修改文件:
  - backend/alembic/versions/2025_09_02_1809-fd315aa4e6b2_initial_database_schema.py
  - backend/app/models/{app,log,tool,user}.py

### 2. 调通bsc-mcp server
- **修复MCP工作目录问题**: backend/app/utils/mcp_client.py:1309
  - 将cwd从MCP_Client改为项目根目录,解决相对路径解析错误
- **修复余额查询API失效**: bsc-mcp/src/functions/fetchBalanceTool.ts
  - 第三方API (termix.ai) 失效,改用viem publicClient直接调用BSC RPC
- **修复私钥格式处理**: bsc-mcp/src/config.ts:107-111
  - 添加0x前缀自动处理,解决privateKeyToAccount报错

### 3. Prompt优化
- 修改intent_service.py和execute_service.py
- 强调"忠实总结",禁止LLM编造或省略工具返回的信息
- 提高max_tokens从200到800,移除字数限制

### 4. 测试用例
- 新增BNB Chain业务测试脚本: run_bnb_tests.py
- 测试结果记录: test_results.md
- **测试统计**:
  - 意图识别命中率: 100% (3/3)
  - 工具选择准确率: 100% (3/3)
  - 参数提取准确率: 100% (3/3)
  - CASE-V-BNB-001 (查询余额): ✅ 成功
  - CASE-V-BNB-002 (BNB兑换USDT): ⚠️ 意图识别成功,执行失败(PancakeSwap API问题)
  - CASE-V-BNB-003 (转账USDT): ⚠️ 意图识别成功,执行失败(余额不足)

## 技术栈
- BSC Mainnet (Chain ID: 56)
- RPC: https://bsc-dataseed.binance.org/
- viem: 直接链上查询
- MCP Server: bsc-mcp (本地构建)

+++ backend/app/models/log.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Text, DateTime, ForeignKey
+from sqlalchemy import Column, Integer, String, Text, DateTime, ForeignKey, BigInteger
 from sqlalchemy.sql import func
 from sqlalchemy.orm import relationship
 from app.utils.db import Base
@@ -7,7 +7,7 @@ class Log(Base):
     """操作日志模型"""
     __tablename__ = "logs"
     
-    id = Column(Integer, primary_key=True, autoincrement=True)
+    id = Column(BigInteger, primary_key=True, autoincrement=True)
     session_id = Column(String(36), ForeignKey("sessions.session_id"), nullable=False)
     step = Column(String(32), nullable=False)  # interpret, execute, confirm等
     status = Column(String(32), nullable=False)  # success, error等
